src/microservices/events/requirements.txt
confluent-kafka
fastapi
kafka-python
uvicorn
